requirements.txt
click==8.0.1
environ==1.0
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
python-dotenv==0.19.0
Werkzeug==2.0.1
